Update hpacucli-check.sh
convert hpacucli-check in ssacli-check and all command "hpacucli" in "ssacli" as the tool from HP has been renamed in the latest version and support old commands and newer controller as well as the older one. This is the tool to use as of today. Some minor changes made as well, adding the report for individual disk in the email, as sometimes SMART of drive could be bad, but the RAID itself is still healthy, it is useful to look these information.